src/vision.py
import cv2
import numpy as np
import mss

class VisionHandler:
    def __init__(self):
        pass
        
    def capture_screen(self, region):
        """
        Capture a specific region of the screen.
        region: {'top': int, 'left': int, 'width': int, 'height': int}
        """
        with mss.mss() as sct:
            screenshot = sct.grab(region)
            img = np.array(screenshot)
            return cv2.cvtColor(img, cv2.COLOR_BGRA2BGR)

    def split_board(self, board_image):
        """
        Split board image into 64 squares.
        Assumption: board_image is cropped exactly to the board edges.
        """
        h, w, _ = board_image.shape
        sq_h, sq_w = h // 8, w // 8
        squares = []
        for r in range(8):
            for c in range(8):
                square = board_image[r*sq_h:(r+1)*sq_h, c*sq_w:(c+1)*sq_w]
                squares.append(square)
        return squares

    def recognize_piece(self, square_image):
        """
        Identify piece in the square.
        Returns: piece char (e.g., 'P', 'k', ' ') or None if empty.
        """
        # TODO: Implement template matching or simple color analysis
        # For now, just return empty
        return None

    def get_fen_from_image(self, board_image):
        """
        Full pipeline to convert board image to FEN string.
        """
        squares = self.split_board(board_image)
        # Process squares...
        # Construct FEN...
        return "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1" # dummy
