requirements.txt
customtkinter
python-chess
opencv-python
pillow
mss
numpy
paramiko
pyglet
